Remove FEC and anything regarding recharter
As I proposed on list.

Note that the final sentence of the first paragraph seems to demand a new paragraph of its own.  That suggests that the first item is actually two distinct items: the definition of a protocol (wire format, reliability, loss detection, congestion control), and the definition of an experimentation platform (version and feature negotiation).  I haven't done that.